Design an abstract class for 2 SendMessageContext
