+++ eventmesh-runtime/src/main/java/org/apache/eventmesh/runtime/core/protocol/tcp/client/session/send/UpStreamMsgContext.java
@@ -66,10 +66,6 @@ public class UpStreamMsgContext extends RetryContext {
         return session;
     }
 
-    public CloudEvent getEvent() {
-        return event;
-    }
-
     public long getCreateTime() {
         return createTime;
     }
